Fix connection message view.
Melange was interpreting 'message' as the sponsor instead of google
and combining the connection and message url components fixed the
issue. Additionally the show_connection redirects had not been updated
to reflect its dissolution into two separate views, so those needed
to be modified to redirect to the corresponding pages for the user.

The messages context parameter had also been removed at some point
so the query to fetch messages needed to be added back so that they
will be displayed to users when viewing the connection.