Move assertion before response data access.
Currently if there is an error in getListResponse and response is
not populated with the expected data attribute, a KeyValue exception
is thrown. Instead this revision places the assertion before this
attribute is accessed which results in a failure and subsequent
output of what the error actually was rather than just an uninformative
exception message.