Modify exception type from Forbidden to BadRequest.
Forbidden (403) is the wrong exception to be thrown here. It isn't
that a user does not have permission to view the data, the only
time list_content will contain a negative ("not found") value is
if the index parameter is missing from the request, Thus this is
a BadRequest error and not a Forbidden one.