Fix indentation issue and allow org admin to view user show page.
Indentation forced the db.put call out of the scope in which it
is intended to be run. Additionally allow users to view their own
connection from both "My Connections" and "Connections for My Orgs"
to make changes from either perspective instead of forcing them to
view it from "Connections for My Orgs" if they are org admins.