Remove role from template sent to user.
Appengine db.get fails when attempting to retrieve a Connection from
the key parameter, but the Connection still exists. Removing this
eliminates the error and also allows org admins the flexibility
to offer one role initially and then change their minds. This
avoids the potentially confusing scenario in which the user views
a Connection expecting the role mentioned in the email and instead
sees another.
viewexpecting one role and see another.