Fix skipped connection test.
The reason this was flaky is because it turns out the order in
which profiles created determines who's logged in. In this case,
the org admin profile was logged in instead of the user which
was preventing emails from being sent to said org admin. Additionally
notify_new_requests is not seeded by the profile utilities, so it is
necessary to set it manually so that the org admin will receive emails.