studentFromKwargs function removed from mutator.
It is replaced by a property with the same name in RequestData class. This property will or will not be removed in the future. After all it is just a wrapper over profile.student_info that just raises an error if the user is not a student.

Also, unit tests for that property added.