Error rendering now respects string marked as safe.
As of Django 1.5, django.utils.html includes a format_html function
that respects whether or not a string has been designated as safe
by django.utils.safestring.mark_safe(). This allows one to include
html in error messages that is known to not contain user input and
render it as is in forms. See issue 1900 for motivation behind
this change.