Upgrade Melange to Django 1.5
We randomly generate a SECRET_KEY in the settings.py module every time
a server is started up.

Melange does not use any of Django's features that rely on SECRET_KEY,
but it is required by Django 1.5. Failing to set this item leads to an
exception when the application is loaded.

By randomly generating the key for every new instance, we guarantee that
Melange does not accidentally rely on any Django feature that depends on
it.

According to Django documentation, it is used to provide cryptographic
signing, and should be set to a unique, unpredictable value.