Some createMentorWithTask calls are replaced by seedProfile and seedTask calls.
The old function does just too much. It silently creates a mentor and assigns a task to him or her. Then, this task is returned. It is not a big problem to explicitly create a mentor first and then seed a task. The flow is clearer, however.

As before, this change only applies to the usages of ProfileHelper which also call createOtherUser.