Changes to testHasTasks.
Another program is not created inside of the test body. The problem is that a task created in line 118 belongs to program of organization which is different than the program to which the task belongs. It is a complex sentence, but the conclusion is that the created environment is not realistic. The test managed to pass. However, it is better to use actual dependencies which occur in production as well.