createWorkSubmission function belonging to GCITaskHelper replaced by a new seedWorkSubmission function.
The latter one is stateless. Also, student is not one of the arguments. It is not arbitrary, as work submissions make sense only for tasks that already have students, i.e. task.student is not None.

Program member of GCITaskHelper class was not used at all by that function. Actually, it was even "shadowed" by task's program.