Initial version of the new organization model that will eventually replace the current one.
The first major change between the current class is that the new model inherits from ndb.Model, so new DB API will be used.

Also, content of help texts is not defined directly in the new class. It is only displayed in context of HTML forms, so it should be defined along with forms.

Right now, this class is very minimalistic and contains only two fields, but new ones will be added shortly.