NDB local context is cleared after each test concludes.
NDB uses three layers to access entities:
- local in-memory cache
- memcache
- datastore

The first one is obviously the fastest. Anyway, it should be cleared between two different test cases, as it is assumed that environment is clean at the begining of each test. Unfortunately, clearing data store stub does not clear the other caches which are used by ndb module, so it has to be done manually.

It is related to issue 9802 of AppEngine platform:
https://code.google.com/p/googleappengine/issues/detail?id=9802