Other fields are returned by getApplicationResponseProperties.
Some survey fields may have other option in which a taker is asked to provide their own answer. It is submitted in a survey form in a field whose id is 'question-id-other'. It should be saved in SurveyResponse model long with question-id. Unfortunately, SurveySchema does not recognize 'question-id-other' as a field, so it is not in field_ids list. This code manually checks whether answers to other options are present in the form data. If such a field is present, it is added to the list of properties to return.