NDB support added to organization() function in Linker class.
It is needed right now, as organizations are getting updated to NDB model. Unfortunately, there is still a lot work to do, so the function now supports both: old db classes and new ndb ones. Support for the old style models will be discarded when all orgs are eventually converted to NDB.