Form renamed to OrgAppForm.
Also, form instances to take org application are not created directly but via a helper function.

The problem with Django forms is that many of their properties are simply defined on class level. Then, when a class is actually initialized, an appropriate metaclass processes them. It is supposed to be convenient for clients (metaclass is responsible for translation between human readable and simple Form classes and Django internal structures), but it has its limitations. One of them is that it is not trivial to modify fields on the fly. This is why the helper functions are added. They take a large, versatile for and costumize it for a particular usage.