SurveyResponseReadOnlyTemplate class added to readonly_template.
It is the first attempt to separate read-only templates from specific model classes and instances. Probably there will be more changes so that the class is more versatile and can be used not only for surveys.

This class is much simpler than the previous SurveyRecordReadOnlyTemplate which inherited from ModelReadOnlyTemplate. It does not need metaclasses and does not require subclassing that class for each model.

Another noticable change is that template_path is passed as argument to the constructor to the class. No longer is it a class variable. This is the direction into which the code should be changing. It makes dependency injection and composition more natural and no subclasses are required.