getPropertyName added to SurveyField class.
It is actually the same as the existing getFieldName function, but is more verbose. Field name is actually not clear in this context, as name may refer to anything. The new function states clearly that the returned string should be used by SurveyResponse to store an answer to the specific question (field).