When a new organization application is submitted, initial org admins will be assigned.
It simplifies (?) the process, as the two users are considered organizations administrators from the moment they submit organization application. Therefore, no checks like: "is this user assigned as the main admin or backup admin for this organization/organization application" are needed anymore. It will be just checked whether a specified user is an organization admin or not.

Additionally, no conversion to org admin will be necessary when organization is accepted or when a person, who took the organization application or who is assigned as the backup admin, creates organization profile.

However, the problem is what to do with organization administrators for organizations which end up rejected? Should they still be considered organization administrators for the program, i.e. is_org_admin set to True? It is questionable, as may affect some priviledges which are normally reserved for 'active' organization administrators or mentors. The main problem I have found so far is the list of all participants definied in participants module. It uses the following query:
    q = GSoCProfile.all()
    q.filter('program', self.data.program)
    q.filter('is_mentor', True)

This may be used by program administrators to obtain list of users who should be sent T-Shirts or other goodies after the program. Should they be sent to the rejected organizations? Of course not. It has to be resolved and the problem will be tackled when profile model is refactored before the next GSoC.