Status property added to new Organization model.
It will determine what is the current status of an organization. It is more similar to the old OrgAppRecord status than to old Organzation status. Anyway, it is fine, as it is all simpler after all. Old 'new' status is not used anymore (as far as I can imagine) and 'invalid' state (also probably not in use except for one GCI organization that was banned one day) can be simulated by REJECTED.