NDB Email model.
It is the same as the previous Email model in app.soc.models.email module. The only difference is that the new one inherits from ndb.Model. Why is it necessary? Email entities are usually created from within a transaction. If it is NDB transaction, it can only handle NDB entities and vice versa.