Mail entity is retrived by a task depending on the style of entity key.
The code is very ugly and will be changed at some point. The problem is that raw string representation is passed to the task. We cannot easily figure out what API does it represent. The only option (at least that I can think of) it to try decoding the strings into keys.