data argument removed from getContext function (a couple of others).
It is a logic function and it should be separated from RequestData which is a part of request handling process. Also, because of that mock RequestData objects had to be created in order to test logic functions.

One can actually notice how much shorter individual tests of setStatus function became.