organization function accepts org_key instead of a full entity.
It was not a problem to obtain the entity so far, but linker will be used to generate URLs for list rows soon. When we speak about new lists, we do not work with entities anymore. Particular rows are repesented by JSONable dictionaries. The only part that can be accessed is: item['columns']['key'] that will evaluate to string representation of organization's key.