PublicOrganizationListPage view added to summerofcode package.
It will list all organizations accepted into the specified program.

Right now, there is only one simple unit test. It is essential to test content of the lists here. I think that some changes will be beneficial when it comes ot testing list content. The main problem is that currently, we have to deal directly with JqgridResponse objects. Their getData method returns a structure which corresponds to the jqgrid protocol. It is the same problem we have with the old lists. Fortunately, I guess it is not very difficult to redesign the concept a little bit so that clients deal with more testable objects.