Unit tests for createOrganizationWithApplication in Summer Of Code-specific context added.
It is tested that it is not possible to create an organization with an invalid ideas page. It is also tested if a valid value is set properly. ideas_page does not make sense for outside of GSoC, so the corresponding tests are added to a new module.

Also, unused imports are removed from the main test_organization module.