Contact model added.
It will be used to represent various electronical contact channels to get in touch with 'a specific' body, for example with an organization. It will probably be also used for programs and maybe for users. All the fields are optional, so it makes sense to use subsets only. For example, feed_url usually does not make sense for individuals.
Anyway, right now such fields are duplicated in different models and are not tested.