Logic for the recently added Contact model.
For now, there is only one function that creates new objects of this model. It is some sort of entry point which does not propagate exceptions if input data is not valid. That would happen, if the constructor was called directly.

Also, simple unit tests are added in this commit.