_default_indexed set to False for SurveyResponse class.
Open end questions are usually accepted in surveys. They cannot be persisted in datastore if their length is longer than 500 characters. This is because all properties of Expando models are indexed by default. On the other hand, expando model properties comes under the same restrictions as text properties. Therefore, long values cannot be indexed.

If we really need it, we may manually endforce indexing of individual properties.