is_veteran property added to Organization model.
It is very similar to the old new_org property which beloned to OrgAppRecord model. It was also a part of GSoCOrganization model. The old name was a bit ambiguous. In the very old days of Melange, new ors were organization which did not created profiles.
is_veteran is clearer as it suggests that it refers to prior participation in the program. Maybe there is a better name, however.