OrgAppUpdatePage class changed to OrgProfileEditPage.
Now that view will be used exclusively to update organization profile. There will be a separate view to update organization application response. This will both simplify the logic and make it easier to outline the pages in a clear way.

The layout will be powered by tabs. One of them will point to a page to edit organization profile, whereas the other one will point to a page showing organization application response.