getProperties functions return None for empty fields.
When form data is submitted to the server, blank fields are represented by empty strings ''. This is a problem, beucase usually it is a sign to specify that no value should be specified for the given property.
It is a more serouis problem for URLs: when the link validator is given an empty string, it says it is not a valid URL and does not accept this value. Of course, we could extend the validator so that empty strings are valid and persisted, but it is clearer when "no value" is specified by None.