OrgProfileCreatePage view added.
It will be an initial view on which a propsective organization administartor creates a new organization profile. When the data is saved successfully, the person will be redirected to Submit Organization Application view on which he or she may answer the questions and complete the registration.

Also, unit tests are included in this revision.