url_user property is used instead of user in GCIProfileShowAdminPage.
This is what is actually expected, as the admin who visits this site does not want to see their own username but the student's one.

It is related to issue 1958.

I am not adding any tests, as this part of code is not easily testable: there should be an easier way to test content of the dictionary which is returned by context() function. However, this module has so many other drawbacks, i.e. functions whose behavior is overriden in subclasses. This set of profile_show modules needs refactoring and I think it should be covered by tests then.