ProgramSelect template is passed template_path in its constructor.
The main reason is that I need such a template which will be rendered in a slightly different way. The only option now is to add another sublass and override templatePath method. Because of this change, it will be possible to simply pass another path to this constructor.

As the default template is used in large majority of places, there is DefaultProgramSelect class which simply takes care of passing in that template.