Initial version of a view to edit organization preferences.
Firstly, it should be defined to what 'preferences' refers. It is a subset of properties of SOCOrganization model that correspond to their participation in the program. Currently, the form contains two fields: minimal and maximal number of slots that the organization requests. It will be exteded with a few extra fields in the future.

Previously, they were set on Organization Profile page, but they do not really belong to profile. Preferences will only be set by accepted organizations.

Also, two very simple unit tests added.