_getPropertiesForFields function pulled out from _OrgProfileForm class.
This function is generic, i.e. there is nothing specific to that particular form. Other parts of code may also want to benefit from it without having to duplicate code.

This function will be likely changed to public and moved to a different module so that its clients can come from all diffrent modules.