events function removed from RedirectHelper.
It is replaced by stateless Linker class. Also, events_link is added to a context only when events page is defined for the specified program. It was also done previosly, but the flow was complicated: events() function tried to check if events_page is defined for a program. If the answer was negative, it would set a local _no_url member. Then, url() function did not return any URL if _no_url was set to True.