A few Key-based utility functions added to prifle module.
These funtions will be used to access information that is hold in keys of entities belonging to Profile model. Format of names of these keys is well-defined and consists of three parts:
<sponsor_id>/<program_id>/<user_id>.
All these individual parts can be obtained just by simple string manipulation. It is convenient, as no actual entity must be retrieved.