EncodedKeyColumn class added to lists module.
We usually try to generate key columns based on their key names or numeric IDs. It is more complex, if an entity has a parent or even a grandparent and so on. Normally, these values are used only by jqGrid. Apparently, it needs a column with unique values in order to work properly.

As the new module for lists is being integrated, one may want to make some computations based on existing ListData. However, that class contains a list of items. Each item is represented by a dict. Valures are usually strings. It might be awkward to play with a typical key representation, like 'sponsor_id/program_id/some_other_name/id'. Some of the parts belong to entity's parent. What is what? Is ID required to be numeric or is it just coincidence? It is much easier, if encoded representation of a key is available, as it can be quickly encoded to an instance of db or ndb Key class.

The problem is that end users may see meaningless strings in Key column. It is not such a big deal, as:
- key column is hidden by default
- the current values are also meaningless. Moreover, it may not be obvious to some users who may try to look for some actual information.

If it ever becomes a serious problem, i.e. angry people yell at us, there are a few possible solutions:
- key column may not be even displayed on the list
- lists module can be modified that some more data is gathered (like encoded Key), but that data is not send back to the client, i.e. it does not belong to any columns.