NDB Organization entity is fetched instead of old organization.
Project entities need to be updated even before the model migrates to NDB. There is org property which is a reference to an organization for a project. Currently, this property points to old DB keys. It has to be changed so that SOCOrganization model is used instead.

If the change above is completed, project.org will stop working, becasuse the underlying key will say that model's name is 'SOCOrganization'. Unfortunately, there is no implementation that inherits from db.Model, so the code will fail.

This change introduces get_value_for_datastore function which will retrieve a raw key. Then, this key is transformed to NDB key, and finally an entity is retrieved. It works, as there is implementation for SOCOrganization that inherits from ndb.Model.

Also, a typo in class name is fixed.