MapReduce script to convert profiles.
Now, it just appends keys for new organizations to the existing lists. So, if there is a list:

mentor_for = [db.Key('GSoCOrganization', 'sponsor_id/program_id/org_id')],

it will be converted to:

mentor_for = [db.Key('GSoCOrganization', 'sponsor_id/program_id/org_id'), db.Key('SOCOrganization', 'sponsor_id/program_id/org_id')].

It is safe, as long as we just check if a particular db.Key is located in that list. It is not safe, if we try to retrive an entity associated with 'SOCOrganization' from it. AppEngine will complain that there is no model implementation for 'SOCOrganization'. It is correct, as there is no DB implementation for 'SOCOrganization' - it is based on NDB.
Anyway, I did not find any place at which we would try to fetch an entity based on those lists. However, it was a quick search.