project_owner field is not set by Mutator.
We can always access owner of a project by calling project.parent(). I think that the main purpose of this function was to save one round trip when project's owner was the current profile. Anyway, it is not a siginificant profit, especially if we take into account that project_owner member was used only once but it was set always. Also, it will be cheaper when profile model migrates to NDB, as it uses in-memory cache for subsequent access to the same entity.