link_id is removed from POST data in testCreateUserNoLinkId.
The essence of this test is that link_id is not present in POST data. However, it was there, and     self.default_props.update({}) did not remove it from the dictionary.

The test used to pass, but the error was revealed by some recent changes (i.e. replacing LinkIDProvider), fortunately.