The code assumes that org property is some models point to 'SOCOrganization'.
The models in question are: GSoCProposal, GSoCProject, GSoCGradingProjectSurveyRecord and GSoCProjectSurveyRecord.

Therefore, it is not possible to do:
entity.org
at this moment. The reason is that the statement above triggers an RPC and retrieves an entity. The problem is that there is no SOCOrganization model in DB namespace.

It is only possible to access a key (by get_value_for_datastore) and then transform that db.Key to ndb.Key and then get the entity. However, it turns out that in many cases we do not need actual entities.