More lenient check in make_value_from_form for ReferenceProperty.
The intention was good: make sure that an entity exists for a supplied key. However, the approach is not correct. This code is not executed within a trasaction, so we cannot conclude that something really exists.
Also, it executes an RPC for each reference property. These fields are usually not provided by users but are set, i.e. manually added to form data by self.cleaned_data['reference_field'] = something, so it is client's responsibility to make sure that an entity exists.

Last but not least, it is a problem for organizations. If DB key is supplied (only DB keys are supported for values of db.ReferenceProperty), but no DB model exists, db.get will raise an error.

Also, we are getting rid of ModelForms.