scoring_disabled property renamed to scoring_enabled in SOCOrganization model.
It is better to keep properties that explicitly enable features than the ones whose purpose is to disable something. It addresses a comment that was left for revision 8762661b522d.

Also, organization administrators are not allowed to post scores, when scoring is not enabled. It concludes the discussion on revision 09d1ba328533.