Next element for NDB data reader is defined by urlsafe representation of cursor.
Previosuly, a string representation of entity's key was used. It was probably added to old lists a long time ago, when App Engine did not offer cursors. Because of that we assumed that keys were ordered by their keys and added another filter: __key__ >= NEXT_KEY.

Using cursors is superior, as this is what cursors have been designed for and there is no need to store another index.

Also, unit tests are updated.