Change download_student_forms to download tax forms for all students.
Previously it downloaded tax forms for just U.S. students, but this
year our Program Administrator wants to look at all student tax forms.

Also this commit clarifies the usage of the "program" argument. It
should be a program path such as "google/gsoc2014" and not simply a
program name.