collection_name changed for various properties in a few models.
It is not used anywhere, as it silently executes a query and was replaced in NDB. Anyhow, proposals name causes a problem, as this error may be thrown:
DuplicatePropertyError: Class GSoCProfile already has property proposals
This new name resolves this problem.

It was probably caused by the fact that reference_class attribute was recently removed from some reference properties because of conversion of organizations.