account_id property added to User model.
There is not much point is storing account as ndb.UserProperty in a model. It consists of two elements and only one of them is immutable. The other one may be changed for a Google Account and it might cause a few problems.

To be more specific, AppEngine's User object has two important parameters by which an account is recognized: it is email and user_id. For a given Google Account, the second element is permanent and will never change, whereas email address may be changed by a user at some point.

Now imagine that ndb.UserProperty is stored in a model. It means that pairs <email, user_id> are persisted in the datastore. It is also possible to query by these values (I had problems today locally, but it was probably my fault). The problem is that they are almost never generated by Melange. Instead, they come from environment.

If a user has a Google Account represented by <X, Y> and creates a Melange account, the application will store exactly that property. Then, if the user changes their email to Z, the app must deal with <Z, Y> and tries to, for example, query by this value. Unfortunately, no results are yield as Melange have not been informed about the change and it still believes that <X, Y> is correct.

This is why it is better to store account_id than account. Unfortunately, it is not feasible to remove account property aright away. There are some old User accounts for which user_id is not know. In that case, we try to query by account. However, account property is not required anymore.