loginNDB function added to profile_utils.
It is similar to login() funtion for DB user entity, but it does not set USER_EMAIL variable.

Why does it sound OK? Because we do not care about it. Fortunately, AppEngine's user class will return a valid user, if USER_ID only is set. Is email important? It might be when we call account.email(), but right now account is not required in User model, so there is a chance that we do not have access to that. If it ever turns out to be important, we can extend the signature of this function so that it accepts email too.