ProfileRegisterAsStudentPage view added.
It will be used by prospective students to register new profiles.

The logic for POST handler is relatively similar to the one for organization members, it has been moved to CreateProfileFormHandler. It is now reused by the new view.

Also, a simple end-to-end unit test is added in this revison.