Connection related views are temporarily removed.
Connection model is about to be converted to NDB. The problem is that all the views quite strongly rely on the fact that the model inherits from db.Model. The main problem is the form: it has to be changed, at lest for GSoC, because model forms do not support NDB models.

In my humble opinion, it will be easier to:
- remove the views
- convert the model (this is actually very easy -> basically replace db with ndb)
- fix the logic
- re-add the views for GSoC
- re-add the views for Code-in when Code-in profiles are converted

Then to do the following:
- keep the views for Code-in
- add NDBConnecion model
- add logic and views related to NDB connection
- convert Code-in profiles
- get rid of old Connection model
- rename NDBConnection to Connection