queryForOrganizationAdmin function removed. queryForOrganizations added instead.
The first function was too generic. It could be applied to any profile, not necessarily one that represents an organization admin. In fact, a number of such cases were convered by tests.

It was fine as long as DB was used, as it allowed explicit empty queries. By that I mean a query whose form is:
SELECT * FROM Model WHERE foo IN []
If the list is emtpy, there is no option the query yields any result. Then would anyone even bother to run that query? This is the idea behind NDB: empty lists are not supported by IN filter - an error is raised. Now, it is the caller's duty to check if the query makes any sense.

Another option would be to make queryForOrganizationAdmin return None for an empty list of orgnizations, but then the caller would probably want to handle that.

This function was used only in one place. Fortunately, it should be invoked only for actual organization administrators.

Unit tests for the old function were removed.
Unit tests for the new function were added.