A method added to RawQueryContentResponseBuilder so as to support NDB models.
There may be no time this year before GSoC to update all existing lists to the new style. It might be necessary to update the existing lists module to work with NDB models. This is the very first step. From now on, it should be possible to use NDB queries for lists.