Drop dependency on ipaddr.
This dependency was added more than three years ago in revision
e12761fa21cf341042b and no one on the team has a solid handle on
why (just that working with Buildout was hard at the time and
explicitly depending on ipaddr fixed some then-immediately acute
problem). Whatever it was, ipaddr is completely unused today.