Modify user and profile keys to reflect new expectations for NDB queries.
Profile keys are dependent on those of the parent, so these are
generated dynamically to prevent changes to the current user from
causing all ndb profile queries to become invalid. This should fix
the issue of local testing running into the "Active profile required..."
error. Also account_id must be assigned to to User.user_id() where User is the
current AppEngine User since that's how ndb Users are queried.