Change all instances of "grade(s)" to "degree(s)".
"Grade" is a wrong translation from Latin based languages to English.

From
http://www.oxforddictionaries.com/definition/english/grade, section "Origin":

"early 16th century: from French, or from Latin gradus 'step'. Originally used
as a unit of measurement of angles (a degree of arc), the term later referred
to degrees of merit or quality."

Also in English it can be confused with Grad, which is a different unit.