Readonly class added.
It will be used as a template to display readonly groups of data, i.e. entities. Something very similar is already used for Organization Applications.

The main reason behind this change is the undergoing conversion to NDB. With old models, we were able to use ModelReadOnlyTemplate. Unfortunately, it does not support NDB and I do not want to play with it, beucase
- it is not straightforward
- it has the same disadvantages as ModelForms provdided by Django/App Engine
- it is too simple to use such big cannons as meta classes.