SOCStudentData is actually created when a new student profile is registered for GSoC.
models was not passed to createStudentData and an instance of StudentData was created. Because of that we were not able to access some fields just by calling profile.student_data.number_of_projects. It applies to fields that are defined in SOCStudentData but not in StudentData.

Also, SOCStudentData properties are tested now.