Work around ndb.StructuredProperty not respecting inheritance
In particular, the field's content is really a SoCStudentData, even
though the Profile entity says it is a StudentData. Since we access the
SoCStudentData through the Profile entity, and it only exposes those
fields that are in StudentData, we can only access the
number_of_projects through the to_dict() method, which has no such
limitation.