Fields from SOCStudentData model are (at least temporarily) moved to StudentData.
The model is used as StructuredProperty in Profile model. It would be appreciated if this property has better support for inheritance. Unfortunately, it turns out (and it is not obvious) that it is fairly limited.

It is possible to supply an instance of a subclass, but it is not possible to access its properties directly. In particular, calling profile.student_data.number_of_project will likely raise # AttributeError: 'StudentData' object has no attribute 'number_of_project'.

An issue to add support for subclasses will be filed, but for now this is a relatively easy and non-invasive solution that we can afford within our time constraints.