Page that displays a list of all survey responses is added.
It will be used by program administrators to see all responses at the same time (and possibly export to CSV). The will be also able to make decisions on which orgs should be accepted into the program.
It is the new version of GSoCOrgAppRecordsList class which will be removed soon.

A smoke unit test is also added.