Status of organizations can be set on the list of org applications.
There is a handler added. The UX is still not perfect: a program host has to double click on the field, then choose a new status, and then click Enter in order to confirm selection and trigger a POST request. It is related to jqgrid lists and there is nothing that can be done at this point to improve that.