Actual BlobInfo objects are created by process_request function.
This is a very nice change. One of the differences between the production and test environment was that the first one generated and persisted BlobInfo entities, whereas the other one returned some datastore.Entity object.

This was caused by strange behavior of BlobstoreServiceStub. In particular, one expects that CreateBlob function creates an actual (possibly mock?) blob object. Instead, it returned just an entity with 'size' property. It was even reported as Issue 9860 for Google App Engine

https://code.google.com/p/googleappengine/issues/detail?id=9860

When one reads the documentation more carefully, it may be noticed that this stub does not intend to be helpful with actual blob management. It is supposed to provide help in managing upload sessions, etc:

"  This stub does not handle the actual creation of blobs, neither the BlobInfo
  in the Datastore nor creation of blob data in the blob_storage.  It does,
  however, assume that another part of the system has created these and
  uses these objects for deletion."

Anyhow, we want to test if blobs are persisted correctly. For the old DB API, we store something called BlobReferenceProperty. Some simple manipulations were probably sufficient to deal with entities returned by CreateBlob function. For NDB, we use BlobKeyProperty and explicit BlobKey instances are required.

Fortunately (I hope), I found that BlobInfo class may work as a wrapper over a datastore entity. That way, we can
- create an entity via CreateBlob
- assign a few extra properties (like filename, etc)
- create BlobInfo manually.

That way, both production and test code deal with BlobInfo rather than different classes.